requirements.txt
streamlit
python-dotenv
openai>=1.30.0
pydantic>=2.0